requirements.txt
django
django_cleanup
djangorestframework
django-cors-headers